refactor(app): 과제 병합 로직 분리
