+++ packages/app/src/features/assignments/index.ts
@@ -1,1 +1,2 @@
 export * from "./hooks";
+export * from "./utils";
