+++ packages/app/src/features/assignments/utils/index.ts
@@ -1,0 +1,1 @@
+export * from "./mergeAssignments.ts";
